multiple associate and group multiple email

+++ database/migrations/2024_08_09_051106_create_email_migration.php
@@ -11,24 +11,26 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('emails', function (Blueprint $table) {
-            $table->id(); 
-            $table->json('toEmail'); 
-            $table->json('ccEmail')->nullable(); 
-            $table->json('bccEmail')->nullable(); 
-            $table->string('fromEmail'); 
-            $table->string('subject')->nullable(); 
-            $table->longText('content')->nullable(); 
-            $table->unsignedBigInteger('userId'); 
-            $table->boolean('isEmailSent')->default(false); 
-            $table->boolean('isDeleted')->default(false); 
-            $table->string('sendEmailFrom')->nullable(); 
-            $table->string('message_id')->nullable(); 
-            $table->timestamps(); 
+        if (!Schema::hasTable('emails')) {
+            Schema::create('emails', function (Blueprint $table) {
+                $table->id(); 
+                $table->json('toEmail'); 
+                $table->json('ccEmail')->nullable(); 
+                $table->json('bccEmail')->nullable(); 
+                $table->string('fromEmail'); 
+                $table->string('subject')->nullable(); 
+                $table->longText('content')->nullable(); 
+                $table->unsignedBigInteger('userId'); 
+                $table->boolean('isEmailSent')->default(false); 
+                $table->boolean('isDeleted')->default(false); 
+                $table->string('sendEmailFrom')->nullable(); 
+                $table->string('message_id')->nullable(); 
+                $table->timestamps(); 
 
-            // Foreign key constraint (optional)
-            $table->foreign('userId')->references('id')->on('users')->onDelete('cascade');
-        });
+                // Foreign key constraint (optional)
+                $table->foreign('userId')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
